# By klesch (336) and others
# Via klesch (202) and others
* commit '4f12198f04d49e781904a5ee5a804c130d54e87d [formerly 05876dbbe05bafccb02f9ce8b04d0a19342e9725]': (3715 commits)
  DF-6857 screen transition fixes
  DF-6857 fix screen switching logic
  [DF-6826] Replace TextBlock::CleanLine to StringUtils::RemoveNonPrintable.
  [DF-6812] Fixes for possible download hang up.
  [DF-6701] Fix review issues
  [DF-6797] fix error in logic.
  [DF-6701] Change interface and work logic of TextLayout
  [DF-6797] Change assert msg to logger msg.
  DF-6784 fix deadlock on filter and create new text field
  DF-6779 review fix
  DF-6779 fix releasing polygon data
  DF-6779 review fix
  DF-6779 helper to release geometry data
  [DF-6776] Fixing webview lazy show, after it was really hidden
  Added c_str() to Format
  [DF-6686] merge fixed
  [DF-6686] added project from development
  [DF-6686] kill files to merge them from development
  [DF-6627] Fixes after review
  [DF-6627] Fixes after review
  ...


Former-commit-id: 7d35a4f2e58063f4d9134fc313ba571fc3d3e078

+++ Sources/Internal/Render/Material/RenderTechnique.h
@@ -100,7 +100,7 @@ public:
 	
 	inline uint32 GetPassCount() const
 	{
-		return renderTechniquePassArray.size();
+		return static_cast<uint32>(renderTechniquePassArray.size());
 	}
 	
 	inline uint16 GetTechniqueId() const
@@ -137,7 +137,6 @@ public:
 protected:
     bool LoadRenderTechnique(const FastName & fastName, RenderTechnique * targetTechnique);
     bool LoadRenderTechniqueFromYamlNode(const YamlNode * rootNode, RenderTechnique * targetTechnique);
-    
     HashMap<FastName, RenderTechnique *> renderTechniqueMap;
 };
